refactoring

+++ app/src/main/java/com/statsup/Value.kt
@@ -3,7 +3,7 @@ package com.statsup
 interface Value {
     fun totalOfYear(position: Int): Double
     fun averageOfYear(position: Int): Double
-    fun ofYear(position: Int): List<Float>
-    fun max(): Float
+    fun ofYear(position: Int): List<Double>
+    fun max(): Double
     fun average(): Double
 }
